Move certain classes to Services
