+++ Src/RandomizerTMF.Logic/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using RandomizerTMF.Logic.Services;
 
 namespace RandomizerTMF.Logic;
