+++ Src/RandomizerTMF.Logic/Services/FilePathManager.cs
@@ -1,4 +1,4 @@
-﻿namespace RandomizerTMF.Logic;
+﻿namespace RandomizerTMF.Logic.Services;
 
 public interface IFilePathManager
 {
