+++ Src/RandomizerTMF.Logic/Session.cs
@@ -1,6 +1,7 @@
 ﻿using GBX.NET.Engines.Game;
 using Microsoft.Extensions.Logging;
 using RandomizerTMF.Logic.Exceptions;
+using RandomizerTMF.Logic.Services;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using TmEssentials;
