+++ Src/RandomizerTMF.Logic/SessionData.cs
@@ -1,5 +1,6 @@
 ﻿using GBX.NET.Engines.Game;
 using Microsoft.Extensions.Logging;
+using RandomizerTMF.Logic.Services;
 using TmEssentials;
 using YamlDotNet.Serialization;
