+++ Src/RandomizerTMF/ViewModels/MainWindowViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using Avalonia.Controls;
 using Avalonia.Media;
 using RandomizerTMF.Logic;
+using RandomizerTMF.Logic.Services;
 using RandomizerTMF.Views;
 using ReactiveUI;
