+++ Src/RandomizerTMF/ViewModels/ModuleWindowViewModelBase.cs
@@ -1,4 +1,5 @@
 ﻿using RandomizerTMF.Logic;
+using RandomizerTMF.Logic.Services;
 
 namespace RandomizerTMF.ViewModels;
