+++ Src/RandomizerTMF/ViewModels/RequestRulesControlViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using RandomizerTMF.Logic;
+using RandomizerTMF.Logic.Services;
 using ReactiveUI;
 using TmEssentials;
