+++ Src/RandomizerTMF/ViewModels/SessionDataViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using GBX.NET.Engines.Game;
 using RandomizerTMF.Logic;
+using RandomizerTMF.Logic.Services;
 using RandomizerTMF.Models;
 using ReactiveUI;
 using System.Collections;
