+++ Src/RandomizerTMF/ViewModels/StatusModuleWindowViewModel.cs
@@ -1,4 +1,4 @@
-﻿using RandomizerTMF.Logic;
+﻿using RandomizerTMF.Logic.Services;
 using ReactiveUI;
 
 namespace RandomizerTMF.ViewModels;
